requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
fastapi==0.128.0
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
kiwisolver==1.4.9
matplotlib==3.10.8
narwhals==2.15.0
numpy==2.4.1
openai==2.15.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
plotly==6.5.2
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pyparsing==3.3.1
pypdf==6.6.0
PyPDF2==3.0.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
wheel==0.45.1
